Modified display judgment of comparison dialog.#126
