+++ jp.go.aist.rtm.rtcbuilder/src/jp/go/aist/rtm/rtcbuilder/ui/editors/DocumentEditorFormPage.java
@@ -1,7 +1,7 @@
 	package jp.go.aist.rtm.rtcbuilder.ui.editors;
 
 import jp.go.aist.rtm.rtcbuilder.generator.param.RtcParam;
-import jp.go.aist.rtm.rtcbuilder.ui.StringUtil;
+import jp.go.aist.rtm.rtcbuilder.util.StringUtil;
 
 import org.eclipse.swt.SWT;
 import org.eclipse.swt.layout.GridData;
